Split shell shot fixtures
